position conversion moved in Utils.kt
